Walking & Gaze script disabling
Created a movement system for the player. Player moves with wasd. While holding w, the player will move towards the direction that the camera is pointing at. Also implemented a system which detects when is player looking at cubes which are rotating. When the player is looking at the cube, the cube stops rotating.